chore: best practices for sharing tests
